feat: earley parsing!
A few things I want to improve:
- Overall code quality
- More idiomatic OrderedSet
- Lex before parse

A lot of parsing algorithms are harder (if not impossible) to implement
without tokenizing. This is a cool opportunity to practice lexing too!

- Proper testing

It'd be cool for parsers to output the left-parse (sequence of productions)
that can generate the input string. That makes it insanely easy to test
and then there's actually some proof that the parser worked.

In conclusion, lots of areas to improve, but exciting nonetheless!!